pyproject.toml
[project]
name = "sim-poc"
version = "0.1.0"
description = "Physics Simulator with FastAPI backend and Elm frontend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "replicate>=0.25.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=5.0.0",
    "pillow>=12.0.0",
    "anthropic>=0.73.0",
    "openai>=2.8.0",
    "structlog>=25.5.0",
    "tenacity>=9.1.2",
    "pydantic-settings>=2.11.0",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "slowapi>=0.1.9",
]
